Unfucked soul heart rendering, mostly
It was using GL11 functions to shift the ENTIRE FUCKING UI for some inexplicable reason. Especially noticeable on a mount, because it was shifting NEI too. Added a config if you need to adjust the height value, as it currently sits in the same spot as the Botania flight bar (and uses some of the code too)
Tweaked lang file. Now with 100% less weeb trash
Removed old load order functions
Cleaned up ConfigHandler