Fix Remedium Talisman breaking other items (#6)
* Fix Remedium Talisman breaking other items

When the Remedium Talisman breaks, it always breaks the item in the first ring slot instead. `BaubleType.AMULET.ordinal()`, whose value is 1, is incorrect. The `BaubleType` Enum is not bauble slot numbers.

The correct slot number for the amulet slot is 0. I don't see any constants to use for the amulet slots.

* Added comment
Added this comment in short to source.